fixup! fixup! refactor: clean up container sandbox image
