+++ packages/sandbox-container/docker-compose.yml
@@ -2,7 +2,7 @@ services:
   autofix-container:
     build:
       context: ../..
-      dockerfile: apps/autofix/src/container-server/Dockerfile
+      dockerfile: packages/sandbox-container/Dockerfile
       args:
         NODE_EXTRA_CA_CERTS: /etc/ssl/certs/Cloudflare_CA.pem
     environment:
